docs(ngClick): describe your change...
I only changed a line so that I could submit this commentary. Important information is missing from the page. There is no discussion about the value provided by the directive and its relationship with the other obvious possible attributesThere. For example, when the result of an 'onclick' html event is false, the navigation indicated by an associated 'href' attribute is not taken. Is the same true for this directive? That was the first question that came to my mind when reading the page.

Just an observation for some constructive improvement in the documentation.